# Conflicts:
#	src/Makefile.am
#	src/Makefile.qt.include
#	src/chainparams.cpp
#	src/init.cpp
#	src/main.cpp
#	src/main.h
#	src/primitives/block.h
#	src/primitives/transaction.cpp
#	src/primitives/transaction.h
#	src/rpcblockchain.cpp
#	src/rpcmining.cpp
#	src/script/interpreter.cpp
#	src/script/interpreter.h
#	src/script/script.cpp
#	src/script/script.h
#	src/script/script_error.h
#	src/script/standard.cpp
#	src/script/standard.h
#	src/wallet.h

+++ src/compressor.h
@@ -114,7 +114,7 @@ public:
     inline void SerializationOp(Stream& s, Operation ser_action, int nType, int nVersion)
     {
         if (!ser_action.ForRead()) {
-            uint64_t nVal = CompressAmount(txout.nValue);
+            uint64_t nVal = CompressAmount((uint64_t)txout.nValue);
             READWRITE(VARINT(nVal));
         } else {
             uint64_t nVal = 0;
